Do not force global `Hash::Merge` behaviour, use a localized behaviour. (#36)
Forcing the global precedence for `Hash::Merge` requires that anyone
using `DBIx::Class::Sims` accept `RIGHT_PRECEDENT` for all of their bare
calls to `merge()`.

Similarly... let's not rely on or expect the global precedence behaviour
for `Hash::Merge` to have the default precedence.  Instead, create our
own instance via `Hash::Merge->new(...)`, and set the precedence we
want.

That way, we know that

(a) we've always got the precedence we want (`RIGHT_PRECEDENT`), and
(b) if someone elses changes the global behaviour, it won't affect us.